Simplify database calls and use higher-level transactions.
Use a scoped_session to make it easy to always get the current database
session with one per thread. In the scheduler and web requests, wrap the
unit of work (a main loop iteration in the scheduler and a
request/response in the Django app) in a transaction so that the whole
thing or none of it will be committed. This also removes the need for
the classlock decorator around all of the Database methods. It also
removes the need for handling commit/rollback behavior in each of those
methods.

This does make us be more aware of when the database is getting
used, by explicitly starting a transaction. In sqlalchemy 2.0+, we can
add `autobegin=False` to the sessionmaker arguments, which will make it
easier to see when database access is being requested (like by using an
attribute of an object that is "expired", i.e. outside of the
transaction it was retrieved in). Without this knowledge, we could be
making unnecessary calls to the database and not know it.

This type of transaction handling also makes it easier to do things like
find and reserve a machine. Improvements to this will be coming in a PR
soon, but for now, we can at least use `with_for_update()` to lock the
row of the machine that is found and, since that single request is not
its own transaction, it can be locked later in the same transaction and
we are guaranteed that no other thread will grab it.